+++ src/dependency_injector/__init__.py
@@ -1,6 +1,6 @@
 """Top-level package."""
 
-__version__ = '4.5.4'
+__version__ = '4.6.0'
 """Version number.
 
 :type: str