+++ src/dependency_injector/ext/flask.py
@@ -9,7 +9,7 @@ from dependency_injector import providers, errors
 
 
 warnings.warn(
-    'Module "dependency_injector.ext.aiohttp" is deprecated since '
+    'Module "dependency_injector.ext.flask" is deprecated since '
     'version 4.0.0. Use "dependency_injector.wiring" module instead.',
     category=DeprecationWarning,
 )