+++ tests/typing/list.py
@@ -27,3 +27,12 @@ provided3: providers.ProvidedInstance = provider3.provided
 attr_getter3: providers.AttributeGetter = provider3.provided.attr
 item_getter3: providers.ItemGetter = provider3.provided['item']
 method_caller3: providers.MethodCaller = provider3.provided.method.call(123, arg=324)
+
+# Test 4: to check the return type with await
+provider4 = providers.List(
+    providers.Factory(object),
+    providers.Factory(object),
+)
+async def _async4() -> None:
+    var1: List[Any] = await provider4()  # type: ignore
+    var2: List[Any] = await provider4.async_()