+++ tests/typing/object.py
@@ -11,3 +11,9 @@ provided2: providers.ProvidedInstance = provider2.provided
 attr_getter2: providers.AttributeGetter = provider2.provided.attr
 item_getter2: providers.ItemGetter = provider2.provided['item']
 method_caller2: providers.MethodCaller = provider2.provided.method.call(123, arg=324)
+
+# Test 3: to check the return type with await
+provider3 = providers.Object(int(3))
+async def _async3() -> None:
+    var1: int = await provider3()  # type: ignore
+    var2: int = await provider3.async_()